requirements.txt
Flask
flask_cors
boto3
python-dotenv
bleach